auto-cite/requirements.txt
manubot==0.5.3
PyYAML==6.0
dict-hash==1.1.26
gscholar